Vertex declaration is now defined as a property in the shader programs

+++ Rendering/DX11/GERenderingShadersDX11.h
@@ -26,10 +26,10 @@ namespace GE { namespace Rendering
       ID3D11InputLayout* dxInputLayout;
       ID3D11VertexShader* dxVertexShader;
 
-      void createInputLayout(const char* pByteCode, uint iByteCodeSize, uint iVertexElements, ID3D11Device1* dxDevice);
+      void createInputLayout(const char* pByteCode, uint iByteCodeSize, uint8_t iVertexElements, ID3D11Device1* dxDevice);
 
    public:
-      VertexShader(const char* Filename, uint VertexElements, const ShaderProgram::PreprocessorMacroList& Macros, ID3D11Device1* DXDevice);
+      VertexShader(const char* Filename, uint8_t VertexElements, const ShaderProgram::PreprocessorMacroList& Macros, ID3D11Device1* DXDevice);
       VertexShader(const char* ByteCode, uint ByteCodeSize, uint VertexElements, ID3D11Device1* DXDevice);
 
       ID3D11InputLayout* getInputLayout() const;
